requirements.txt
numpy==1.24.2
pandas==1.3.4
nibabel==5.1.0
scipy==1.9.1
natsort==8.4.0
einops==0.7.0
matplotlib==3.5.0
tensorboardX==2.6.2.2
scikit-learn==0.24.0
torchio==0.19.6
